Remove default cost strategy and associated tests
Unfortunately either the Fixed-exp or the Fixed-ln function seem to fail badly
when big numbers are either the result (exp) or the input (ln) and the amount of
fractional bits is low. The default cost strategy is able to calculate the result
without an overflow, but the results quality is miserable.
Example vector: q=[500.0, 500.0], fee=0.0107734, type=FixedU128/I128<U33> -> about 10% off.
This does not happen with the optimized cost strategy, which is now used in any case.