Check that all pool join/exit's raise on zero
`util::pool` raises if either `pool_amount` or `asset_amount` is zero.
This is an unfortunate result of the abstraction of `pool_join` and
`pool_exit` into `util::pool`, but it does protect the user from paying
for something they get nothing for.

The other four functions only raise an error if what the joining/exiting
user pays is zero. So this protects the LPs, but not the interacting
user.