Check if report matches market type using method
We move some of the free functionality of `prediction-markets` into the
`Market` primitive. This makes the functionality available in other
pallets which use the `Market` primitive, such as `authorized`.

In `authorized`, we now check that the report actually matches the
market type.

We also use `test-case` for testing now. This should be very useful,
especially when doing unit tests for primitives.